Move `support/helpers` to base test directory and change the route alias to be tests
